Override the HTTP header containing the user
For some reason, NGINX considers request headers containing underscores as invalid and drops them silently. So, we can use the FLASK_AUTHNZ_USER_HEADER environment variable to indicate the header. But be careful with the whitelist in this case. We typically want to squash any headers coming in thru the request; we have the logic for REMOTE_USER but a new header may not be addressed from a security perspective.