Support for self-contained apps.
We can pass in the name of the database that contains the main roles collection. For self-contained applications, this is the same database that stores the other business objects.

Bear in mind, if we used this, we will not get instrument level permissions as the instruments are stored in the "site" database.